requirements.txt
openslide-python
opencv-python
scikit-image
numpy
bitarray
bitarray-hardbyte
python-dotenv